refactor(file structure): move store files to own feature store dirs
