+++ src/app/fitness-schedule/filter/filter.component.spec.ts
@@ -4,7 +4,7 @@ import { FilterComponent } from './filter.component';
 import { SelectFilterComponent } from './select-filter/select-filter.component';
 import { FormsModule } from '@angular/forms';
 import { MappingService } from '../services/mapping.service';
-import { FilterActions } from '../store/filter.actions';
+import { FilterActions } from '../store/filter/filter.actions';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { TimeFilterComponent } from './time-filter/time-filter.component';
 import { SfsMaterialModule } from '../../material/sfs-material.module';
