+++ src/app/fitness-schedule/fitness-schedule.module.ts
@@ -6,21 +6,21 @@ import { ScheduleComponent } from './schedule/schedule.component';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { SortClassesService } from './services/sort-classes.service';
 import { FilterClassesService } from './services/filter-classes.service';
-import { ScheduleEpics } from './store/schedule.epics';
-import { ScheduleActions } from './store/schedule.actions';
+import { ScheduleEpics } from './store/schedule/schedule.epics';
+import { ScheduleActions } from './store/schedule/schedule.actions';
 import { FilterComponent } from './filter/filter.component';
 import { FormsModule } from '@angular/forms';
 import { SelectFilterComponent } from './filter/select-filter/select-filter.component';
 import { MappingService } from './services/mapping.service';
-import { FilterActions } from './store/filter.actions';
+import { FilterActions } from './store/filter/filter.actions';
 import { OrderDaysPipe } from './pipes/order-days.pipe';
 import { RouterModule, Routes } from '@angular/router';
 import { SettingsComponent } from './settings/settings.component';
-import { SettingsActions } from './store/settings.actions';
+import { SettingsActions } from './store/settings/settings.actions';
 import { SfsMaterialModule } from '../material/sfs-material.module';
 import { ScheduleParserService } from './services/schedule-parser.service';
 import { TimeFilterComponent } from './filter/time-filter/time-filter.component';
-import { FavoriteActions } from './store/favorite.actions';
+import { FavoriteActions } from './store/favorites/favorite.actions';
 import { FilteredScheduleComponent } from './filtered-schedule/filtered-schedule.component';
 import { FavoritesScheduleComponent } from './favorites-schedule/favorites-schedule.component';
 import { DayColumnComponent } from './day-column/day-column.component';
