+++ src/app/fitness-schedule/store/favorites/favorite-state.ts
@@ -1,4 +1,4 @@
-import { FitnessClass } from './fitness-class';
+import { FitnessClass } from '../../interfaces/fitness-class';
 export interface FavoriteState {
   workouts: FitnessClass[];
 }
