+++ src/app/fitness-schedule/store/filter/filter-state.ts
@@ -1,6 +1,6 @@
-import { Day } from '../enums/day.enum';
-import { Gym } from '../enums/gym.enum';
-import { Language } from '../enums/language.enum';
+import { Day } from '../../enums/day.enum';
+import { Gym } from '../../enums/gym.enum';
+import { Language } from '../../enums/language.enum';
 export interface FilterState {
   minStartTime?: number;
   maxEndTime?: number;
