+++ src/app/store/root.epics.ts
@@ -4,7 +4,7 @@
 import { Injectable } from '@angular/core';
 import 'rxjs/add/operator/catch';
 import 'rxjs/add/operator/map';
-import { ScheduleEpics } from '../fitness-schedule/store/schedule.epics';
+import { ScheduleEpics } from '../fitness-schedule/store/schedule/schedule.epics';
 import { createEpicMiddleware } from 'redux-observable';
 import { PushNotificationEpics } from '../push-notification/push-notification.epics';
