+++ src/app/store/root.reducers.ts
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux';
-import { scheduleReducer } from '../fitness-schedule/store/schedule.reducers';
-import { filterReducer } from '../fitness-schedule/store/filter.reducers';
-import { settingsReducer } from '../fitness-schedule/store/settings.reducers';
-import { favoriteReducer } from '../fitness-schedule/store/favorite.reducers';
+import { scheduleReducer } from '../fitness-schedule/store/schedule/schedule.reducers';
+import { filterReducer } from '../fitness-schedule/store/filter/filter.reducers';
+import { settingsReducer } from '../fitness-schedule/store/settings/settings.reducers';
+import { favoriteReducer } from '../fitness-schedule/store/favorites/favorite.reducers';
 import { pushNotificationReducer } from '../push-notification/push-notification.reducers';
 /**
  * Created by Kim on 02.04.2017.
